pyproject.toml
[project]
name = "agent-deployment-context-eng"
version = "0.1.0"
description = "Multi-agent system for financial data analysis using Databricks Genie"
requires-python = ">=3.12"
dependencies = [
    "mlflow[databricks]==3.2.0",
    "databricks-langchain==0.6.0",
    "databricks-agents==1.4.0",
    "langgraph==0.5.4",
    "pydantic<2.12.0",
    "nest-asyncio==1.6.0",
    "databricks-cli",
    "numpy>=1.26.4",
    "ipykernel>=6.30.1",
]

[project.optional-dependencies]
dev = []

[tool.uv]
dev-dependencies = []
